Quick fix for downloading OIFITS files with curl.
The key on which to sort the results of an ADQL query (ORDER BY clause)
have to be in the selection list. This is not the case currently (for
example on the one for downloading with curl).

For the time being, the order clause from the source query is cleared.